ARREGLANDO PETICION FOTO

+++ SICAPL/repositorios/repositorio_administrador.inc.php
@@ -21,40 +21,36 @@ class Repositorio_administrador {
                 $email = $administrador->getEmail();
                 $fecha = $administrador->getFecha();
                 $administradorExistente = self::obtener_administrador($conexion, $codigo_administrador);
-                $EmailExistente = self::obtener_email($conexion, $email);
+                $emailExistente = self::obtener_email($conexion, $email);
                 $duiExistente = self::verifica_dui($conexion, $dui);
 
+                if ($administradorExistente->getCodigo_administrador() == "" || $emailExistente->getEmail() == '' || $duiExistente->getDui() == '') {
+
+                    $sql = 'INSERT INTO administradores(codigo_administrador,pasword,nivel,nombre,apellido,sexo,dui,estado,observacion,foto,email,fecha)'
+                            . ' values (:codigo_administrador,:pasword,:nivel,:nombre,:apellido,:sexo,:dui,:estado,:observacion,:foto,:email,:fecha)';
+                    ///estos son alias para que PDO pueda trabajar 
+                    $sentencia = $conexion->prepare($sql);
+
+                    $sentencia->bindParam(':codigo_administrador', $codigo_administrador, PDO::PARAM_STR);
+                    $sentencia->bindParam(':pasword', password_hash($pasword, PASSWORD_DEFAULT), PDO::PARAM_STR);
+                    $sentencia->bindParam(':nivel', $nivel, PDO::PARAM_INT);
+                    $sentencia->bindParam(':nombre', $nombre, PDO::PARAM_STR);
+                    $sentencia->bindParam(':apellido', $apellido, PDO::PARAM_STR);
+                    $sentencia->bindParam(':sexo', $sexo, PDO::PARAM_BOOL);
+                    $sentencia->bindParam(':dui', $dui, PDO::PARAM_STR);
+                    $sentencia->bindParam(':estado', $estado, PDO::PARAM_STR);
+                    $sentencia->bindParam(':observacion', $observacion, PDO::PARAM_STR);
+                    $sentencia->bindParam(':email', $email, PDO::PARAM_STR);
+                    $sentencia->bindParam(':foto', $foto, PDO::PARAM_STR);
+                    $sentencia->bindParam(':fecha', $fecha, PDO::PARAM_STR);
+
+                    $administrador_insertado = $sentencia->execute();
+                    $mensaje = 'Se registro como administrador a ' . $nombre . ' ' . $apellido;
+
+                    self::insertar_bitacora($conexion, $mensaje);
 
 
-
-                if ($duiExistente->getCodigo_administrador() == "") {
-                    if ($EmailExistente->getEmail() == "") {
-                        if ($duiExistente->getDui() == "") {
-                            $sql = 'INSERT INTO administradores(codigo_administrador,pasword,nivel,nombre,apellido,sexo,dui,estado,observacion,foto,email,fecha)'
-                                    . ' values (:codigo_administrador,:pasword,:nivel,:nombre,:apellido,:sexo,:dui,:estado,:observacion,:foto,:email,:fecha)';
-                            ///estos son alias para que PDO pueda trabajar 
-                            $sentencia = $conexion->prepare($sql);
-
-                            $sentencia->bindParam(':codigo_administrador', $codigo_administrador, PDO::PARAM_STR);
-                            $sentencia->bindParam(':pasword', password_hash($pasword, PASSWORD_DEFAULT), PDO::PARAM_STR);
-                            $sentencia->bindParam(':nivel', $nivel, PDO::PARAM_INT);
-                            $sentencia->bindParam(':nombre', $nombre, PDO::PARAM_STR);
-                            $sentencia->bindParam(':apellido', $apellido, PDO::PARAM_STR);
-                            $sentencia->bindParam(':sexo', $sexo, PDO::PARAM_BOOL);
-                            $sentencia->bindParam(':dui', $dui, PDO::PARAM_STR);
-                            $sentencia->bindParam(':estado', $estado, PDO::PARAM_STR);
-                            $sentencia->bindParam(':observacion', $observacion, PDO::PARAM_STR);
-                            $sentencia->bindParam(':email', $email, PDO::PARAM_STR);
-                            $sentencia->bindParam(':foto', $foto, PDO::PARAM_STR);
-                            $sentencia->bindParam(':fecha', $fecha, PDO::PARAM_STR);
-
-                            $administrador_insertado = $sentencia->execute();
-                            $mensaje = 'Se registro como administrador a ' . $nombre . ' ' . $apellido;
-
-                            self::insertar_bitacora($conexion, $mensaje);
-
-
-                            echo '<script>swal({
+                    echo '<script>swal({
                     title: "Exito",
                     text: "El registro ha sido Guardado!",
                     type: "success",
@@ -65,31 +61,9 @@ class Repositorio_administrador {
                     location.href="inicio_seguridad.php";
                     
                 });</script>';
-                        }else{
-                          echo '<script>'
-                        . 'swal("Cuidado!", "El dui que introdujo ya esta en uso, favor introdusca otro", "warning");'
-                        . '$("#idNombre").val("' . $nombre . '"); $("#idApellido").val("' . $apellido . '");'
-                        . '$("#idUser").val("' . $codigo_administrador . '"); $("#idDui").val("' . $dui . '");'
-                        . '$("#idFecha").val("' . $fecha . '"); $("#idEmail").val("' . $email . '");'
-                        . 'if ("' . $nivel . '" == "0") {$("#idRoot").attr("checked", "checked");} else {$("#idAdministrador").attr("checked", "checked");}'
-                        . 'if ("' . $sexo . '" == "Masculino") {$("#idHombre").attr("checked", "checked");} else {$("#idMujer").attr("checked", "checked");}'
-                        . '$("#idListarAdmnistrador").removeClass("active");  $("#idRegistroAdministrador").addClass("active"); '
-                        . '$("#idPass1").val("' . $pasword . '"); $("#idPass2").val("' . $pasword . '");  </script>';  
-                        }
-                    } else {
-                        echo '<script>'
-                        . 'swal("Cuidado!", "El correo que introdujo ya esta en uso, favor introdusca otro", "warning");'
-                        . '$("#idNombre").val("' . $nombre . '"); $("#idApellido").val("' . $apellido . '");'
-                        . '$("#idUser").val("' . $codigo_administrador . '"); $("#idDui").val("' . $dui . '");'
-                        . '$("#idFecha").val("' . $fecha . '"); $("#idEmail").val("' . $email . '");'
-                        . 'if ("' . $nivel . '" == "0") {$("#idRoot").attr("checked", "checked");} else {$("#idAdministrador").attr("checked", "checked");}'
-                        . 'if ("' . $sexo . '" == "Masculino") {$("#idHombre").attr("checked", "checked");} else {$("#idMujer").attr("checked", "checked");}'
-                        . '$("#idListarAdmnistrador").removeClass("active");  $("#idRegistroAdministrador").addClass("active"); '
-                        . '$("#idPass1").val("' . $pasword . '"); $("#idPass2").val("' . $pasword . '");  </script>';
-                    }
-                } else {
+                } else if ($administradorExistente->getCodigo_administrador() != '') {
                     echo '<script>'
-                    . 'swal("Cuidado!", "El nombre de usuario que introdujo ya esta en uso, favor introdusca otro", "warning");'
+                    . 'swal("Cuidado!", "El Usuario que introdujo ya esta en uso, favor introdusca otro", "warning");'
                     . '$("#idNombre").val("' . $nombre . '"); $("#idApellido").val("' . $apellido . '");'
                     . '$("#idUser").val("' . $codigo_administrador . '"); $("#idDui").val("' . $dui . '");'
                     . '$("#idFecha").val("' . $fecha . '"); $("#idEmail").val("' . $email . '");'
@@ -97,6 +71,26 @@ class Repositorio_administrador {
                     . 'if ("' . $sexo . '" == "Masculino") {$("#idHombre").attr("checked", "checked");} else {$("#idMujer").attr("checked", "checked");}'
                     . '$("#idListarAdmnistrador").removeClass("active");  $("#idRegistroAdministrador").addClass("active"); '
                     . '$("#idPass1").val("' . $pasword . '"); $("#idPass2").val("' . $pasword . '");  </script>';
+                }else if ($duiExistente->getDui()!= '') {
+                      echo '<script>'
+                        . 'swal("Cuidado!", "El dui que introdujo ya esta en uso, favor introdusca otro", "warning");'
+                        . '$("#idNombre").val("' . $nombre . '"); $("#idApellido").val("' . $apellido . '");'
+                        . '$("#idUser").val("' . $codigo_administrador . '"); $("#idDui").val("' . $dui . '");'
+                        . '$("#idFecha").val("' . $fecha . '"); $("#idEmail").val("' . $email . '");'
+                        . 'if ("' . $nivel . '" == "0") {$("#idRoot").attr("checked", "checked");} else {$("#idAdministrador").attr("checked", "checked");}'
+                        . 'if ("' . $sexo . '" == "Masculino") {$("#idHombre").attr("checked", "checked");} else {$("#idMujer").attr("checked", "checked");}'
+                        . '$("#idListarAdmnistrador").removeClass("active");  $("#idRegistroAdministrador").addClass("active"); '
+                        . '$("#idPass1").val("' . $pasword . '"); $("#idPass2").val("' . $pasword . '");  </script>'; 
+                }else if ($emailExistente->getEmail() != '') {
+                      echo '<script>'
+                        . 'swal("Cuidado!", "El Correo que introdujo ya esta en uso, favor introdusca otro", "warning");'
+                        . '$("#idNombre").val("' . $nombre . '"); $("#idApellido").val("' . $apellido . '");'
+                        . '$("#idUser").val("' . $codigo_administrador . '"); $("#idDui").val("' . $dui . '");'
+                        . '$("#idFecha").val("' . $fecha . '"); $("#idEmail").val("' . $email . '");'
+                        . 'if ("' . $nivel . '" == "0") {$("#idRoot").attr("checked", "checked");} else {$("#idAdministrador").attr("checked", "checked");}'
+                        . 'if ("' . $sexo . '" == "Masculino") {$("#idHombre").attr("checked", "checked");} else {$("#idMujer").attr("checked", "checked");}'
+                        . '$("#idListarAdmnistrador").removeClass("active");  $("#idRegistroAdministrador").addClass("active"); '
+                        . '$("#idPass1").val("' . $pasword . '"); $("#idPass2").val("' . $pasword . '");  </script>'; 
                 }
             } catch (PDOException $ex) {
                 //echo '<script>swal("Advertencia!", "Favor revisar los datos e intentar nuevamente", "warning");</script>';
@@ -116,7 +110,7 @@ class Repositorio_administrador {
         }
     }
 
-    public static function obtener_administrador($conexion, $codigo_administrador) {
+       public static function obtener_administrador($conexion, $codigo_administrador) {
         $administrador = new Administrador();
         if (isset($conexion)) {
             try {
@@ -129,6 +123,7 @@ class Repositorio_administrador {
                     $administrador->setEmail($row["email"]);
                     $administrador->setNombre($row["nombre"]);
                     $administrador->setApellido($row["apellido"]);
+                    $administrador->setDui($row["dui"]);
                 }
             } catch (PDOException $ex) {
                 print 'ERROR: ' . $ex->getMessage();
